requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
langchain_google_genai
langchain-openai